Move some of the padding creation funcs to where makes more sense
